chore: move filtering functionality to search utils
