+++ app/component/StopsNearYouSearch.js
@@ -1,36 +1,11 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { pure } from 'recompose';
-import compact from 'lodash/compact';
 import DTAutoSuggest from '@digitransit-component/digitransit-component-autosuggest';
-import { filterStopsAndStationsByMode } from '@digitransit-search-util/digitransit-search-util-query-utils';
-import { getGTFSId } from '@digitransit-search-util/digitransit-search-util-suggestion-to-location';
+import { filterSearchResultsByMode } from '@digitransit-search-util/digitransit-search-util-query-utils';
 import withSearchContext from './WithSearchContext';
 
 function StopsNearYouSearch({ mode, breakpoint }) {
-  const filterSearchResultsByMode = (results, type) => {
-    switch (type) {
-      case 'Routes':
-        return results;
-      case 'Stops': {
-        const gtfsIds = results.map(x => {
-          const gtfsId = x.properties.gtfsId
-            ? x.properties.gtfsId
-            : getGTFSId({ id: x.properties.id });
-          if (gtfsId) {
-            return {
-              gtfsId,
-              ...x,
-            };
-          }
-          return null;
-        });
-        return filterStopsAndStationsByMode(compact(gtfsIds), mode);
-      }
-      default:
-        return results;
-    }
-  };
   const DTAutoSuggestWithSearchContext = withSearchContext(DTAutoSuggest);
   const isMobile = breakpoint !== 'large';
   const transportMode = `route-${mode}`;
